src/main/java/seedu/address/storage/XmlAddressBookStorage.java
package seedu.address.storage;

import seedu.address.commons.core.LogsCenter;
import seedu.address.commons.exceptions.DataConversionException;
import seedu.address.model.AddressBook;

import java.io.File;
import java.io.FileNotFoundException;
import java.util.Optional;
import java.util.logging.Logger;

/**
 * A class to access AddressBook data stored as an xml file on the hard disk.
 */
public class XmlAddressBookStorage {

    private static final Logger logger = LogsCenter.getLogger(XmlAddressBookStorage.class);

    public Optional<AddressBook> readAddressBook(String filePath) {
        assert filePath != null;
        return null;
    }
}
